requirements.txt
# Playwright and test framework
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.3
pytest-asyncio==0.21.1
pytest-html==4.1.1
pytest-ordering==0.6
allure-pytest==2.13.2

# OpenAI for self-healing
openai==1.3.7

# Configuration and utilities
jproperties==2.1.1
python-dotenv==1.0.0

# Additional utilities
Pillow==10.1.0

# MFA/TOTP support
pyotp==2.9.0

